+++ src/app/info/[...slug]/docker-nextjs-mysql.tsx
@@ -1,0 +1,5 @@
+import DockerNextjsMysql from "./docker-nextjs-mysql.mdx";
+
+export default function Page() {
+  return <DockerNextjsMysql />;
+}